backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-levenshtein==0.23.0
regex==2023.10.3
nltk==3.8.1
numpy>=1.26.0
